SkypeWeb: TRouter errors handling. Other fixes.
git-svn-id: http://svn.miranda-ng.org/main/trunk@13440 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/SkypeWeb/src/requests/capabilities.h
@@ -29,7 +29,8 @@ public:
 			<< CHAR_VALUE("Content-Type", "application/json; charset=UTF-8")
 			<< FORMAT_VALUE("RegistrationToken", "registrationToken=%s", regToken);
 
-		int bitness = 32;
+		short bitness = 32;
+
 		char compName[MAX_COMPUTERNAME_LENGTH + 1];
 		DWORD size = SIZEOF(compName);
 		GetComputerNameA(compName, &size);
